\name{fastSOM-package}
\Rdversion{1.1}
\alias{fastSOM-package}
\alias{fastSOM}
\docType{package}
\title{
  Fast Calculation of Spillover Measures
}
\description{
  This package comprises various functions for computing spillover measures, especially spillover tables and spillover indices
  as proposed by Diebold and Yilmaz (2009) as well as their estimated and exact average, minimal, and maximal values.
}
\details{
  \tabular{ll}{
    Package: \tab fastSOM\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2016-07-20\cr
    License: \tab GPL (>=2)
  }
}
\author{
  Stefan Kloessner (\email{S.Kloessner@mx.uni-saarland.de}), \cr
  with contributions by Sven Wagner (\email{sven.wagner@mx.uni-saarland.de})
}
\references{
  [1] Diebold, F. X. and Yilmaz, K. (2009):  \href{http://onlinelibrary.wiley.com/doi/10.1111/j.1468-0297.2008.02208.x/pdf}{Measuring financial asset return and volatitliy spillovers, 
                                             with application to global equity markets}, 
                                             Economic Journal 199(534): 158-171.
  
  [2] Kloessner, S. and Wagner, S. (2012):  \href{http://onlinelibrary.wiley.com/doi/10.1002/jae.2366/pdf}{Exploring All VAR Orderings for Calculating Spillovers? Yes, We Can! - 
											A Note on Diebold and Yilmaz (2009)}, Journal of Applied Econometrics 29(1): 172-179
}
\keyword{ package }
